+++ rotkehlchen/externalapis/etherscan_like.py
@@ -1,30 +1,48 @@
+import abc
 import logging
 import operator
-from abc import ABC, abstractmethod
+from abc import ABC
 from collections.abc import Iterator
+from enum import Enum, auto
+from http import HTTPStatus
+from json.decoder import JSONDecodeError
 from typing import TYPE_CHECKING, Any, Final, Literal, overload
+
+import gevent
+import requests
 
 from rotkehlchen.chain.evm.constants import GENESIS_HASH, ZERO_ADDRESS
 from rotkehlchen.chain.structures import TimestampOrBlockRange
 from rotkehlchen.db.constants import TX_DECODED
 from rotkehlchen.db.evmtx import DBEvmTx
 from rotkehlchen.db.history_events import DBHistoryEvents
+from rotkehlchen.db.settings import CachedSettings
+from rotkehlchen.errors.misc import RemoteError
 from rotkehlchen.errors.serialization import DeserializationError
+from rotkehlchen.externalapis.interface import ExternalServiceWithApiKey
+from rotkehlchen.externalapis.utils import get_earliest_ts
 from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.serialization.deserialize import (
     deserialize_evm_transaction,
+    deserialize_int_from_str,
     deserialize_timestamp,
 )
 from rotkehlchen.types import (
     SUPPORTED_CHAIN_IDS,
+    ApiKey,
+    ChainID,
     ChecksumEvmAddress,
     EvmInternalTransaction,
     EvmTransaction,
     EVMTxHash,
+    ExternalService,
     Location,
     Timestamp,
     deserialize_evm_tx_hash,
 )
+from rotkehlchen.utils.misc import set_user_agent
+from rotkehlchen.utils.network import create_session
+from rotkehlchen.utils.serialization import jsonloads_dict
 
 if TYPE_CHECKING:
     from rotkehlchen.db.dbhandler import DBHandler
@@ -44,35 +62,144 @@ def _hashes_tuple_to_list(hashes: set[tuple[EVMTxHash, Timestamp]]) -> list[EVMT
     return [x[0] for x in sorted(hashes, key=operator.itemgetter(1))]
 
 
-class EtherscanLikeApi(ABC):
+class HasChainActivity(Enum):
+    """
+    Classify the type of transaction first found in blockscout/etherscan.
+    TRANSACTIONS means that the endpoint for transactions/internal transactions
+    had entries, TOKENS means that the tokens endpoint had entries, BALANCE means
+    that the address has a non-zero native asset balance and NONE means that no
+    activity was found."""
+    TRANSACTIONS = auto()
+    TOKENS = auto()
+    BALANCE = auto()
+    NONE = auto()
+
+
+class EtherscanLikeApi(ExternalServiceWithApiKey, ABC):
     """Base class for any APIs similar to etherscan."""
 
     def __init__(
             self,
             database: 'DBHandler',
             msg_aggregator: 'MessagesAggregator',
+            service_name: ExternalService,
             pagination_limit: int,
-            name: str,
+            default_api_key: ApiKey,
     ) -> None:
-        self.db = database
+        super().__init__(database=database, service_name=service_name)
         self.msg_aggregator = msg_aggregator
+        self.session = create_session()
+        set_user_agent(self.session)
+        self.default_api_key = default_api_key
         self.pagination_limit = pagination_limit
-        self.name = name
+        self.name = service_name.name.capitalize()
+
+    @staticmethod
+    @abc.abstractmethod
+    def _get_url(chain_id: SUPPORTED_CHAIN_IDS) -> str:
+        """Get the API URL for the given chain. Override in subclasses for different endpoints.
+        May raise UnsupportedChain if the service does not support the given chain.
+        """
+
+    @staticmethod
+    @abc.abstractmethod
+    def _build_query_params(
+            module: str,
+            action: str,
+            api_key: ApiKey,
+            chain_id: SUPPORTED_CHAIN_IDS,
+    ) -> dict[str, str]:
+        """Build parameters for API requests. Override in subclasses for different formats."""
 
     @overload
-    @abstractmethod
     def _query(
             self,
             chain_id: SUPPORTED_CHAIN_IDS,
-            module: Literal['account'],
-            action: Literal['txlistinternal', 'txlist', 'tokentx'],
+            module: str,
+            action: Literal[
+                'balancemulti',
+                'txlist',
+                'txlistinternal',
+                'tokentx',
+                'getLogs',
+                'txsBeaconWithdrawal',
+            ],
             options: dict[str, Any] | None = None,
             timeout: tuple[int, int] | None = None,
     ) -> list[dict[str, Any]]:
         ...
 
     @overload
-    @abstractmethod
+    def _query(
+            self,
+            chain_id: SUPPORTED_CHAIN_IDS,
+            module: str,
+            action: Literal[
+                'eth_getTransactionReceipt',
+                'eth_getTransactionByHash',
+            ],
+            options: dict[str, Any] | None = None,
+            timeout: tuple[int, int] | None = None,
+    ) -> dict[str, Any] | None:
+        ...
+
+    @overload
+    def _query(
+            self,
+            chain_id: SUPPORTED_CHAIN_IDS,
+            module: str,
+            action: Literal[
+                'getcontractcreation',
+            ],
+            options: dict[str, Any] | None = None,
+            timeout: tuple[int, int] | None = None,
+    ) -> list[dict[str, Any]] | None:
+        ...
+
+    @overload
+    def _query(
+            self,
+            chain_id: SUPPORTED_CHAIN_IDS,
+            module: str,
+            action: Literal[
+                'getabi',
+            ],
+            options: dict[str, Any] | None = None,
+            timeout: tuple[int, int] | None = None,
+    ) -> str | None:
+        ...
+
+    @overload
+    def _query(
+            self,
+            chain_id: SUPPORTED_CHAIN_IDS,
+            module: str,
+            action: Literal[
+                'eth_getBlockByNumber',
+            ],
+            options: dict[str, Any] | None = None,
+            timeout: tuple[int, int] | None = None,
+    ) -> dict[str, Any]:
+        ...
+
+    @overload
+    def _query(
+            self,
+            chain_id: SUPPORTED_CHAIN_IDS,
+            module: str,
+            action: Literal[
+                'balance',
+                'tokenbalance',
+                'eth_blockNumber',
+                'eth_getCode',
+                'eth_call',
+                'getblocknobytime',
+            ],
+            options: dict[str, Any] | None = None,
+            timeout: tuple[int, int] | None = None,
+    ) -> str:
+        ...
+
     def _query(
             self,
             chain_id: SUPPORTED_CHAIN_IDS,
@@ -81,27 +208,138 @@ class EtherscanLikeApi(ABC):
             options: dict[str, Any] | None = None,
             timeout: tuple[int, int] | None = None,
     ) -> list[dict[str, Any]] | str | int | dict[str, Any] | None:
-        ...
+        """Queries etherscan-like service
 
-    @abstractmethod
-    def _query(
-            self,
-            chain_id: SUPPORTED_CHAIN_IDS,
-            module: str,
-            action: str,
-            options: dict[str, Any] | None = None,
-            timeout: tuple[int, int] | None = None,
-    ) -> list[dict[str, Any]] | str | int | dict[str, Any] | None:
-        ...
+        None is a valid result for this function when the requested information doesn't exist.
+        Happens when asking for the code of a contract, transaction by hash, receipt...
 
-    @abstractmethod
-    def get_blocknumber_by_time(
-            self,
-            chain_id: SUPPORTED_CHAIN_IDS,
-            ts: Timestamp,
-            closest: Literal['before', 'after'] = 'before',
-    ) -> int:
-        ...
+        May raise:
+        - RemoteError if there are any problems with reaching the service or if
+        an unexpected response is returned. Also in the case of exhausting the backoff time.
+        """
+        result = None
+        if (api_key := self._get_api_key()) is None and self.default_api_key is not None:
+            api_key = self.default_api_key
+            log.debug(f'Using default {self.name} key')
+
+        params = self._build_query_params(
+            module=module,
+            action=action,
+            api_key=api_key,
+            chain_id=chain_id,
+        )
+        if options:
+            params.update(options)
+
+        backoff = 1
+        cached_settings = CachedSettings()
+        timeout = timeout or cached_settings.get_timeout_tuple()
+        backoff_limit = cached_settings.get_query_retry_limit()
+        api_url = self._get_url(chain_id=chain_id)
+        response = None
+        while backoff < backoff_limit:
+            log.debug(f'Querying {self.name} for {chain_id}: {api_url} with params: {params}')
+            try:
+                response = self.session.get(url=api_url, params=params, timeout=timeout)
+            except requests.exceptions.RequestException as e:
+                raise RemoteError(f'{self.name} API request failed due to {e!s}') from e
+
+            if response.status_code == HTTPStatus.TOO_MANY_REQUESTS:
+                if backoff >= backoff_limit:
+                    raise RemoteError(
+                        f'Getting {self.name} too many requests error '
+                        f'even after we incrementally backed off while querying {chain_id}',
+                    )
+
+                log.debug(
+                    f'Got too many requests error from {chain_id} {self.name}. Will '
+                    f'backoff for {backoff} seconds.',
+                )
+                gevent.sleep(backoff)
+                backoff *= 2
+                continue
+
+            if response.status_code != 200:
+                raise RemoteError(
+                    f'{self.name} API request {response.url} failed '
+                    f'with HTTP status code {response.status_code} and text '
+                    f'{response.text}',
+                )
+
+            try:
+                json_ret = jsonloads_dict(response.text)
+            except JSONDecodeError as e:
+                raise RemoteError(
+                    f'{self.name} API request {response.url} returned invalid '
+                    f'JSON response: {response.text}',
+                ) from e
+
+            try:
+                if (result := json_ret.get('result')) is None:
+                    if action in {'eth_getTransactionByHash', 'eth_getTransactionReceipt', 'getcontractcreation'}:  # noqa: E501
+                        return None
+
+                    raise RemoteError(
+                        f'Unexpected format of {self.name} response for request {response.url}. '
+                        f'Missing a result in response. Response was: {response.text}',
+                    )
+
+                # successful proxy calls do not include a status
+                status = int(json_ret.get('status', 1))
+
+                if status != 1:
+                    if status == 0:
+                        if result == 'Contract source code not verified':
+                            return None
+                        if json_ret.get('message', '') == 'NOTOK':
+                            if result.startswith((
+                                'Max calls per sec rate',  # short-term rate limit (5 seconds)
+                                'Max rate limit reached',  # different variant of the message above
+                                'Free API access is temporarily unavailable',  # free tier apikey blocked during high load periods  # noqa: E501
+                            )):
+                                log.debug(
+                                    f'Got response: {response.text} from {self.name} while '
+                                    f'querying chain {chain_id}. Will backoff for {backoff} seconds.',  # noqa: E501
+                                )
+                                gevent.sleep(backoff)
+                                backoff *= 2
+                                continue
+
+                            elif result.startswith('Max daily'):
+                                raise RemoteError(f'{self.name} max daily rate limit reached.')
+
+                    transaction_endpoint_and_none_found = (
+                        status == 0 and
+                        json_ret['message'] == 'No transactions found' and
+                        action in {'txlist', 'txlistinternal', 'tokentx', 'txsBeaconWithdrawal'}
+                    )
+                    logs_endpoint_and_none_found = (
+                        status == 0 and
+                        json_ret['message'] == 'No records found' and
+                        'getLogs' in action
+                    )
+                    if transaction_endpoint_and_none_found or logs_endpoint_and_none_found:
+                        return []
+
+                    # else
+                    raise RemoteError(f'{chain_id} {self.name} returned error response: {json_ret}')  # noqa: E501
+            except KeyError as e:
+                raise RemoteError(
+                    f'Unexpected format of {chain_id} {self.name} response for request {response.url}. '  # noqa: E501
+                    f'Missing key entry for {e!s}. Response was: {response.text}',
+                ) from e
+
+            # success, break out of the loop and return result
+            return result
+
+        # will only run if we get out of the loop due to backoff limit
+        assert response is not None, 'This loop always runs at least once and response is not None'
+        msg = (
+            f'{chain_id} {self.name} API request to {response.url} failed due to backing'
+            ' off for more than the backoff limit'
+        )
+        log.error(msg)
+        raise RemoteError(msg)
 
     def _process_timestamp_or_blockrange(
             self,
@@ -331,3 +569,68 @@ class EtherscanLikeApi(ABC):
             options = new_options
 
         yield _hashes_tuple_to_list(hashes)
+
+    def get_blocknumber_by_time(
+            self,
+            chain_id: SUPPORTED_CHAIN_IDS,
+            ts: Timestamp,
+            closest: Literal['before', 'after'] = 'before',
+    ) -> int:
+        """Performs the etherscan like api call to get the blocknumber by a specific timestamp
+
+        May raise:
+        - RemoteError if there are any problems with reaching the service or if
+        an unexpected response is returned
+        """
+        # set per-chain earliest timestamps that can be turned to blocks. Never returns block 0
+        if ts < get_earliest_ts(chain_id):
+            return 0  # etherscan does not handle timestamps close to genesis well
+
+        options = {'timestamp': ts, 'closest': closest}
+        result = self._query(
+            chain_id=chain_id,
+            module='block',
+            action='getblocknobytime',
+            options=options,
+        )
+        try:
+            number = deserialize_int_from_str(result, f'{self.name} getblocknobytime')
+        except DeserializationError as e:
+            raise RemoteError(
+                f'Could not read blocknumber from {self.name} for {chain_id}  '
+                f'getblocknobytime result {result}',
+            ) from e
+
+        return number
+
+    def has_activity(
+            self,
+            chain_id: SUPPORTED_CHAIN_IDS,
+            account: ChecksumEvmAddress,
+    ) -> HasChainActivity:
+        """Queries native asset balance, transactions, internal_txs and tokentx for an address
+        with limit=1 just to quickly determine if the account has had any activity in the chain.
+        We make a distinction between transactions and ERC20 transfers since ERC20
+        are often spammed. If there was no activity at all we return the enum value
+        NONE.
+        """
+        options = {'address': str(account), 'page': 1, 'offset': 1}
+        result = self._query(chain_id=chain_id, module='account', action='txlist', options=options)
+        if len(result) != 0:
+            return HasChainActivity.TRANSACTIONS
+        result = self._query(chain_id=chain_id, module='account', action='txlistinternal', options=options)  # noqa: E501
+        if len(result) != 0:
+            return HasChainActivity.TRANSACTIONS
+        result = self._query(chain_id=chain_id, module='account', action='tokentx', options=options)  # noqa: E501
+        if len(result) != 0:
+            return HasChainActivity.TOKENS
+        if chain_id in {ChainID.ETHEREUM, ChainID.GNOSIS}:
+            balance = self._query(
+                chain_id=chain_id,
+                module='account',
+                action='balance',
+                options={'address': account},
+            )
+            if int(balance) != 0:
+                return HasChainActivity.BALANCE
+        return HasChainActivity.NONE