+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`-` rotki will now support blockscout and routescan as fallback indexers for retrieving transactions when etherscan fails or does not support a given chain.
 * :bug:`-` Users can now click a location in the manual balance summary on the dashboard to filter by that location.
 * :bug:`-` rotki will now shows the protocol icon from DS Proxy correctly.
 * :bug:`11005` Fix an issue where adding a new manual balance with a custom asset/price unnecessarily fetches prices for all assets.